% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmst.R
\name{cmst}
\alias{cmst}
\title{CMST Tests for genotype and phenotypes with covariates}
\usage{
cmst(pheno, geno, resp_names, addcov, intcov, method = c("par", "non.par",
  "joint", "all"), penalty = c("bic", "aic", "both"), verbose = FALSE)
}
\arguments{
\item{pheno}{data frame with phenotypes and covariates}

\item{geno}{data frame with QTL genotypes}

\item{resp_names}{response names}

\item{addcov, intcov}{additive and interactive covariate names for responses}

\item{method}{method for CMST test (parametric, non-parametric, joint or all three); can provide more than one value.}

\item{penalty}{type of information criteria penalty (for BIC or AIC)}

\item{verbose}{verbose output if \code{TRUE}}
}
\description{
Run CMST tests with one QTL and 2 or more phenotypes.
The \code{geno} is the driver for phenotype responses, with covariates acting on responses.
The \code{resp_names}, \code{addcov} and \code{intcov} names must all be valid names in the \code{pheno} data frame.
}
