% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMSTtests.R
\name{CMSTtests}
\alias{CMSTtests}
\title{CMST Tests for cross object}
\usage{
CMSTtests(cross, pheno1, pheno2, Q.chr, Q.pos, addcov1 = NULL,
  addcov2 = NULL, intcov1 = NULL, intcov2 = NULL, method = c("all",
  "par", "non.par", "joint"), penalty = c("both", "bic", "aic"),
  verbose = FALSE)
}
\arguments{
\item{cross}{object of class cross; see \code{\link[qtl]{read.cross}}.}

\item{pheno1}{first phenotype name}

\item{pheno2}{other phenotype names to test with \code{pheno1}}

\item{Q.chr, Q.pos}{chromosome name (character) and position (in map units).}

\item{addcov1, addcov2, intcov1, intcov2}{additive and interactive covariates for \code{pheno1} and \code{pheno2}.}

\item{method}{method for CMST test (parametric, non-parametric, joint or all three); can provide more than one value.}

\item{penalty}{type of information criteria penalty (for BIC or AIC)}

\item{verbose}{verbose output if \code{TRUE}}
}
\description{
Set up and run CMST tests for cross object. See \code{\link{cmst}}.
}
