% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_scatter.R
\name{med_scatter}
\alias{med_scatter}
\title{Scatter plot for mediator and target}
\usage{
med_scatter(driver, target, mediator, kinship, cov_tar, cov_med, fitFunction,
  sdp = 32, allele = TRUE)
}
\arguments{
\item{driver}{vector or matrix with driver values}

\item{target}{vector or 1-column matrix with target values}

\item{mediator}{vector or 1-column matrix with mediator values}

\item{kinship}{optional kinship matrix among individuals}

\item{cov_tar}{optional covariates for target}

\item{cov_med}{optional covariates for mediator}

\item{fitFunction}{function to fit models with driver, target and mediator}

\item{sdp}{SNP distribution pattern for plot colors}

\item{allele}{Driver has alleles if \code{TRUE}, otherwise allele pairs.}
}
\description{
Scatter plot for mediator and target
}
