% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate1_test.R
\name{mediate1_test}
\alias{mediate1_test}
\title{Develop mediation models from driver, target and mediator}
\usage{
mediate1_test(driver, target, mediator, fitFunction, kinship = NULL,
  cov_tar = NULL, cov_med = NULL, annotation, test = c("wilc", "binom",
  "joint", "norm"), pos = NULL, lod_threshold = 5.5, ...)
}
\arguments{
\item{driver}{vector or matrix with driver values}

\item{target}{vector or 1-column matrix with target values}

\item{mediator}{matrix with mediator values by column}

\item{fitFunction}{function to fit models with driver, target and mediator}

\item{kinship}{optional kinship matrix among individuals}

\item{cov_tar}{optional covariates for target}

\item{cov_med}{optional covariates for mediator}

\item{annotation}{Table with annotation, with \code{id} 
agreeing with column names of \code{mediator}.}

\item{test}{}

\item{pos}{Position of driver.}

\item{lod_threshold}{LOD threshold to include mediator.}

\item{...}{additional parameters}
}
\description{
Develop mediation models from driver, target and mediator
}
