% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_pair.R
\name{mediation_pair}
\alias{mediation_pair}
\alias{ggplot_mediation_pair}
\title{Scatter plot for mediator and target}
\usage{
mediation_pair(target, mediator, driver, covar_tar, covar_med, kinship,
  fitFunction, sdp, allele = TRUE)

ggplot_mediation_pair(x, ..., type = c("by_mediator", "by_target",
  "driver_offset", "driver"), tname = "target", mname = "mediator",
  dname = "driver", centerline = 0, main = tname)
}
\arguments{
\item{target}{vector or 1-column matrix with target values}

\item{mediator}{vector or 1-column matrix with mediator values}

\item{driver}{vector or matrix with driver values}

\item{covar_tar}{optional covariates for target}

\item{covar_med}{optional covariates for mediator}

\item{kinship}{optional kinship matrix among individuals}

\item{fitFunction}{function to fit models with driver, target and mediator}

\item{sdp}{SNP distribution pattern for plot colors}

\item{allele}{Driver has alleles if \code{TRUE}, otherwise allele pairs.}

\item{x}{object of class \code{mediation_pair}}

\item{\dots}{additional parameters for plotting}

\item{type}{type of plot: one of \code{("by_mediator", "by_target", "driver_offset", "driver")}}

\item{tname}{target name (default \code{"target"})}

\item{mname}{mediator name (default \code{"mediator"})}

\item{dname}{driver name (default \code{"driver"})}

\item{centerline}{horizontal line at value (default = \code{0}); set to \code{NA} for no line or \code{NULL} for mean}

\item{main}{main title (defautl \code{tname})}
}
\description{
Scatter plot for mediator and target
}
