% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_test.R
\name{mediation_test}
\alias{mediation_test}
\title{Develop mediation models from driver, target and mediator}
\usage{
mediation_test(target, mediator, driver, annotation, covar_tar = NULL,
  covar_med = NULL, kinship = NULL, driver_med = NULL,
  test = c("wilcoxon", "binomial", "joint", "normal"), pos = NULL,
  fitFunction = fitQtl2, data_type = c("phenotype", "expression"), ...)
}
\arguments{
\item{target}{vector or 1-column matrix with target values}

\item{mediator}{matrix of mediators}

\item{driver}{vector or matrix with driver values}

\item{annotation}{optional annotation data frame for mediators}

\item{covar_tar}{optional covariates for target}

\item{covar_med}{optional covariates for mediator}

\item{kinship}{optional kinship matrix among individuals}

\item{driver_med}{optional driver matrix for mediators}

\item{test}{Type of CMST test.}

\item{pos}{Position of driver.}

\item{fitFunction}{function to fit models with driver, target and mediator}

\item{data_type}{Type of mediator data.}

\item{...}{additional parameters}
}
\description{
Develop mediation models from driver, target and mediator
}
