% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl2cmst.R
\name{qtl2cmst}
\alias{qtl2cmst}
\title{Set up data for cmst function}
\usage{
qtl2cmst(cross, pheno1, pheno2, Q.chr, Q.pos, addcov1 = NULL,
  addcov2 = NULL, intcov1 = NULL, intcov2 = NULL)
}
\arguments{
\item{cross}{object of class cross; see \code{\link[qtl]{read.cross}}.}

\item{pheno1}{first phenotype name}

\item{pheno2}{other phenotype names to test with \code{pheno1}}

\item{Q.chr, Q.pos}{chromosome name (character) and position (in map units).}

\item{addcov1, addcov2, intcov1, intcov2}{additive and interactive covariates for \code{pheno1} and \code{pheno2}.}
}
\description{
Set up R/qtl cross object and identifiers for use with \code{\link{cmst}}.
The \code{pheno} and \code{cov} arguments must be valid names of \code{cross$pheno}.
The \code{Q.chr} must be a valid chromosome identifier for \code{cross}.
}
