% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quatrad_CMST.R
\name{quatrad_CMST}
\alias{quatrad_CMST}
\title{Causal Model Selection Tests for mediation quatrads}
\usage{
quatrad_CMST(models, test = c("wilc", "binom", "joint", "norm"),
  threshold = 0.1)
}
\arguments{
\item{models}{Object from \code{\link{mediationModels}}}

\item{test}{Type of CMST test to perform}

\item{threshold}{Threshold for secondary test of best model across quatrads}
}
\description{
Causal Model Selection Tests for mediation quatrads
}
