% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimCrossCausal.R
\name{SimCrossCausal}
\alias{SimCrossCausal}
\title{Simulate cross object for CMST example}
\usage{
SimCrossCausal(n.ind, len, n.mar, beta, add.eff, dom.eff, sig2.1 = 1,
  sig2.2 = 1, eq.spacing = FALSE, cross.type = c("bc", "f2"),
  normalize = FALSE)
}
\arguments{
\item{n.ind}{number of individuals}

\item{len}{length(s) of chromosome(s)}

\item{n.mar}{markers per chromosome}

\item{beta}{causal effect(s) (slope(s)) of y1 on the other phenotypes}

\item{add.eff, dom.eff}{additive and dominance effects}

\item{eq.spacing}{use equal marker spacing if \code{TRUE}}

\item{cross.type}{cross type (only "bc" and "f2" implemented)}

\item{normalize}{use normal scores if \code{TRUE}}

\item{sig.1, sig.2}{residual variances for y1 and other responses}
}
\description{
This simulates a \code{cross} object with 3 phenotypes where y1 has a causal effect
on both y2 and y3. See vignette for details.
+ \epsilon$ and $y_3 = 0.5 \, y_1 + \epsilon$). The length of beta controls the number of phenotypes to be simulated.
}
