% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate1_test.R
\name{mediate1_test}
\alias{mediate1_test}
\title{Develop mediation models from driver, target and mediator}
\usage{
mediate1_test(mediator, driver, target, kinship = NULL, cov_tar = NULL,
  cov_med = NULL, driver_med = NULL, test = c("wilc", "binom", "joint",
  "norm"), pos = NULL, fitFunction = qtl2scan::fit1,
  data_type = c("expression", "phenotype"), ...)
}
\arguments{
\item{mediator}{List with mediators and annotation.}

\item{driver}{vector or matrix with driver values}

\item{target}{vector or 1-column matrix with target values}

\item{kinship}{optional kinship matrix among individuals}

\item{cov_tar}{optional covariates for target}

\item{cov_med}{optional covariates for mediator}

\item{test}{Type of CMST test.}

\item{pos}{Position of driver.}

\item{fitFunction}{function to fit models with driver, target and mediator}

\item{data_type}{Type of mediator data.}

\item{...}{additional parameters}
}
\description{
Develop mediation models from driver, target and mediator
}
